+++ parlai/core/agents.py
@@ -41,7 +41,8 @@ This module also provides a utility method:
 """
 
 from parlai.core.build_data import modelzoo_path
-from parlai.utils.misc import warn_once, load_opt_file
+from parlai.utils.misc import warn_once
+from parlai.core.opt import load_opt_file
 from .metrics import Metrics, aggregate_metrics
 import copy
 import importlib
