+++ parlai/core/params.py
@@ -18,7 +18,7 @@ import git
 from parlai.core.agents import get_agent_module, get_task_module
 from parlai.core.build_data import modelzoo_path
 from parlai.tasks.tasks import ids_to_tasks
-from parlai.utils.misc import Opt, load_opt_file
+from parlai.core.opt import Opt, load_opt_file
 
 from typing import List, Optional
